+++ src/components/Profile/MyPost/index.jsx
@@ -1,25 +1,11 @@
 import { createRef } from "react";
 
 import Post from "./Post";
-import {
-  updateNewPostTextActionCreator,
-  addPostActionCreator,
-} from "./../../../redux/profile-reducer";
 
 import styles from "./MyPost.module.css";
 
 const MyPost = (props) => {
   const areaRef = createRef();
-
-  const onPostHandler = () => {
-    const text = areaRef.current.value;
-    props.dispatch(updateNewPostTextActionCreator(text));
-  };
-
-  const addPostHandler = () => {
-    props.dispatch(addPostActionCreator());
-    areaRef.current.value = "";
-  };
 
   const PostElements = ({ messageList }) =>
     messageList.map((post, index) => (
@@ -31,12 +17,12 @@ const MyPost = (props) => {
       <h3>My posts</h3>
       <div>
         <textarea
-          onChange={onPostHandler}
+          onChange={props.onPostHandler}
           ref={areaRef}
           placeholder="add your post here"
         />
       </div>
-      <button onClick={addPostHandler}>Add Post</button>
+      <button onClick={props.addPostHandler}>Add Post</button>
       <PostElements messageList={props.messageList} />
     </div>
   );
