+++ src/components/Profile/index.jsx
@@ -2,20 +2,16 @@ import React from "react";
 import MyInfo from "./MyInfo";
 
 import MyPost from "./MyPost";
+import MyPostContainer from "./MyPost/index container";
 
 import styles from "./Profile.module.css";
 
 const Profile = (props) => {
-  const { messageList, addPost, updateNewPostText } = props;
   return (
     <div>
       <div className={styles.content}>
         <MyInfo />
-        <MyPost
-          addPost={addPost}
-          messageList={messageList}
-          dispatch={props.dispatch}
-        />
+        <MyPostContainer store={props.store} messageList={props.messageList} />
       </div>
     </div>
   );
